moved out material-ui for materialize so style not duplicated, added carousel feature, made App Main component
