+++ package.json
@@ -6,10 +6,10 @@
   "dependencies": {
     "babel-preset-es2015": "^6.24.1",
     "gh-pages": "^1.1.0",
-    "material-ui": "^0.20.0",
     "prismjs": "^1.11.0",
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
+    "react-materialize": "^1.1.2",
     "variadic.js": "0.0.21"
   },
   "devDependencies": {
