+++ src/components/FuncParams.js
@@ -1,11 +1,11 @@
 import React, {Component} from 'react';
-import {TextField} from 'material-ui';
+import {Input} from 'react-materialize';
 
 class FuncParams extends Component {
   constructor(props) {
     super(props);
     this.state = {
-      value: 0
+      value: ""
     };
 
     this.handleInputChange = this.handleInputChange.bind(this);
@@ -20,8 +20,8 @@ class FuncParams extends Component {
     e.preventDefault();
     let params = this.props.params.slice();
     params.push(Number(this.state.value));
-    this.textField.select();
     this.props.onParamsChange(params);
+    this.setState({value: ""})
   }
 
   render() {
@@ -29,13 +29,13 @@ class FuncParams extends Component {
 
     return(
       <form onSubmit={this.handleParamSubmit}>
-        <TextField
-          ref={(input) => {this.textField = input;}}
+        <Input
+          value={this.state.value}
           id={`${funcName}-input`}
           type="number"
           step="0.01"
-          hintText="Enter number and hit RETURN"
-          errorText={error}
+          placeholder="Enter number and hit RETURN"
+          error={error}
           onChange={this.handleInputChange}
         />
       </form>
