+++ src/containers/App.js
@@ -2,8 +2,9 @@ import React, { PureComponent } from 'react';
 import FuncCard from "../components/FuncCard";
 import NavBar from "../components/NavBar";
 import variadic from 'variadic.js';
+import {Carousel} from "react-materialize";
 
-class CardList extends PureComponent {
+class App extends PureComponent {
   constructor(props) {
     super(props);
     this.state = {
@@ -34,25 +35,28 @@ class CardList extends PureComponent {
     } = this.state;
 
     let funcCards = funcNames.map((funcName, i) => (
+      <div key={i} style={{width: "400px"}}>
         <FuncCard
-          key={i}
           funcName={funcName}
           func={variadic[funcName]}
           subtitle={subtitle}
           onClickHandler={this.handleClick}
           currentFunc={currentFunc}
         />
+      </div>
     ));
 
     return (
       <div>
         <NavBar />
         <div style={flexContainer}>
-          {funcCards}
+          <Carousel>
+            {funcCards}
+          </Carousel>
         </div>
       </div>
     );
   }
 }
 
-export default CardList;
+export default App;
